+++ lib/lightning/policies/project_users.ex
@@ -74,7 +74,8 @@ defmodule Lightning.Policies.ProjectUsers do
              :edit_data_retention,
              :add_project_user,
              :remove_project_user,
-             :edit_run_settings
+             :edit_run_settings,
+             :create_collection
            ],
       do: project_user.role in [:owner, :admin]
 
@@ -86,8 +87,7 @@ defmodule Lightning.Policies.ProjectUsers do
              :run_workflow,
              :provision_project,
              :create_project_credential,
-             :initiate_github_sync,
-             :create_collection
+             :initiate_github_sync
            ],
       do: project_user.role in [:owner, :admin, :editor]